requirements.txt
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
altair>=5.0.0,<6.0.0
tenacity>=8.2.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
psutil>=5.8.0,<6.0.0
plotly>=5.18.0,<6.0.0
